+++ src/main/java/ru/bellintegrator/practice/view/UserFilterView.java
@@ -16,7 +16,7 @@ public class UserFilterView {
 
     public String possition;
 
-    public Long officeId;
+    public Integer officeId;
 
     public Integer docCode;
 
@@ -57,11 +57,11 @@ public class UserFilterView {
         this.possition = possition;
     }
 
-    public Long getOfficeId() {
+    public Integer getOfficeId() {
         return officeId;
     }
 
-    public void setOfficeId(Long officeId) {
+    public void setOfficeId(Integer officeId) {
         this.officeId = officeId;
     }
